:recycle: Extract job_message.go from job_subscription.go
